% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuralODE.R
\name{backward}
\alias{backward}
\title{Backward pass of the Neural ODE}
\usage{
backward(model, tsteps, outputs, output_gradients = NULL)
}
\arguments{
\item{model}{A keras neural network that defines the Neural ODE.}

\item{tsteps}{A vector of each time step upon which the Neural ODE is solved to get to the final solution.}

\item{outputs}{The tensor outputs of the forward pass of the Neural ODE.}

\item{output_gradients}{The tensor gradients of the loss function.}
}
\value{
The model input at the last time step.

The gradient of loss with respect to the inputs for use with the Adjoint Method.

The gradients of loss the neural ODE.
}
\description{
Backward pass of the Neural ODE
}
\examples{
# example code

}
