% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuralODE.R
\name{backward_dynamics}
\alias{backward_dynamics}
\title{Internal function to solve the backwards dynamics of the Neural ODE}
\usage{
backward_dynamics(state, model)
}
\arguments{
\item{state}{The current state of the differential equation}

\item{model}{The neural network that defines the Neural ODE.}
}
\value{
Returns a list of the number 1, the new backwards state of the differential equation and the gradients calculated for the network.
}
\description{
Internal function to solve the backwards dynamics of the Neural ODE
}
\keyword{internal}
