% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eulermethod.R
\name{euler_step}
\alias{euler_step}
\title{A function to employ the Euler Method to solve an ODE.}
\usage{
euler_step(func, dt, state)
}
\arguments{
\item{func}{The derivative function.}

\item{dt}{The time step for the Euler solver.}

\item{state}{A list that defines the current state of the ODE.}
}
\value{
A list that describes the updated state of the ODE.
}
\description{
A function to employ the Euler Method to solve an ODE.
}
