% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eulermethod.R
\name{euler_update}
\alias{euler_update}
\title{A Euler method state updater.}
\usage{
euler_update(h_list, dh_list, dt)
}
\arguments{
\item{h_list}{The initial state of the ODE.}

\item{dh_list}{description}

\item{dt}{The time step to update the initial state with.}
}
\value{
The updated state of the ODE.
}
\description{
A Euler method state updater.
}
