% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rk4.R
\name{rk4_step}
\alias{rk4_step}
\title{Runge Kutta solver for ordinary differential equations}
\usage{
rk4_step(func, dt, state)
}
\arguments{
\item{func}{The function to be numerically integrated.}

\item{dt}{Time step.}

\item{state}{A list describing the state of the function, with the first element
being 1, and the second being a tensor that represents state}
}
\value{
A list containing a new time and the numerical integration of of the
function across the time step to the new time.
}
\description{
Runge Kutta solver for ordinary differential equations
}
\examples{
# example code


}
