% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuralODE.R
\name{rk4_step_backwards}
\alias{rk4_step_backwards}
\title{Custom internal RK4 solver for solving the backward pass of the Neural ODE.}
\usage{
rk4_step_backwards(backward_dynamics, dt, state, model)
}
\arguments{
\item{backward_dynamics}{The backward dynamics function for the Neural ODE.}

\item{dt}{The time step to solve the ODE on.}

\item{state}{The current state of the differential equation.}

\item{model}{The neural network that defines the Neural ODE.}
}
\value{
An output list with the updated backwards state.
}
\description{
Custom internal RK4 solver for solving the backward pass of the Neural ODE.
}
\keyword{internal}
