% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuralODE.R
\name{forward}
\alias{forward}
\title{Forward pass of the Neural ODE network}
\usage{
forward(model, inputs, tsteps, return_states = FALSE)
}
\arguments{
\item{model}{A keras neural network that defines the Neural ODE.}

\item{inputs}{Matrix or vector inputs to the neural network.}

\item{tsteps}{A vector of each time step upon which the Neural ODE is solved to get to the final solution.}

\item{return_states}{A boolean which dictates whether the intermediary states between the input and the final solution are returned.}
}
\value{
solution of the forward pass of Neural ODE
}
\description{
Forward pass of the Neural ODE network
}
